Java-TwitchBot v2 - Refactored
I've refactored nearly all of the bot's code. I've had to remove a lot
of the features it had, but at the same time made the bot much more
versatile and able to used by anyone.

You can now add as many bot accounts as you want in the account.cfg
file. As long as each one is in a different channel, there shouldn't be
any issues, but if they are, the only real difference is that they won't
know they're all in the same channel, and so they'll repeat eachother.

Modules are on the way too. I already have a framework planned out, I
just need to think of a proper way to implement it.

Please, report any bugs so they can be fixed.