Added the first "Complex" module
As a test of the module framework, I created a small TicTacToe game for
the bots. I discovered a few things about modules while I was at it.
Will update readme. For now, I'll leave the TicTacToeModule in the
source so that others can see how I made it and how to make their own.